refactor: refactoring of middleware.test.js to repective folders
