+++ test/middleware.test.js
@@ -22,8 +22,6 @@ import webpackErrorConfig from './fixtures/webpack.error.config';
 import webpackMultiErrorConfig from './fixtures/webpack.array.error.config';
 import webpackWarningConfig from './fixtures/webpack.warning.config';
 import webpackMultiWarningConfig from './fixtures/webpack.array.warning.config';
-import webpackOneErrorOneWarningOneSuccessConfig from './fixtures/webpack.array.one-error-one-warning-one-success';
-import webpackOneErrorOneWarningOneSuccessWithNamesConfig from './fixtures/webpack.array.one-error-one-warning-one-success-with-names';
 
 describe('middleware', () => {
   let instance;
@@ -59,600 +57,6 @@ describe('middleware', () => {
       }
     });
   }
-
-  describe('basic', () => {
-    describe('should respect the "stats" option with the "false" value from the configuration', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler({ ...webpackConfig, stats: false });
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request to the bundle file', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            expect(getLogsPlugin.logs).toMatchSnapshot();
-
-            return done();
-          });
-      });
-    });
-
-    describe('should respect the "stats" option with the "none" value from the configuration', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler({ ...webpackConfig, stats: 'none' });
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request to the bundle file', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            expect(getLogsPlugin.logs).toMatchSnapshot();
-
-            return done();
-          });
-      });
-    });
-
-    describe('should respect the "stats" option with the "minimal" value from the configuration', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler({ ...webpackConfig, stats: 'minimal' });
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request to the bundle file', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            expect(getLogsPlugin.logs).toMatchSnapshot();
-
-            return done();
-          });
-      });
-    });
-
-    describe('should respect the "stats" option in multi-compiler mode', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackOneErrorOneWarningOneSuccessConfig);
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" requests to bundles file', (done) => {
-        request(app)
-          .get('/static-one/bundle.js')
-          .expect(200, (firstError) => {
-            if (firstError) {
-              return done(firstError);
-            }
-
-            return request(app)
-              .get('/static-two/bundle.js')
-              .expect(200, (secondError) => {
-                if (secondError) {
-                  return done(secondError);
-                }
-
-                return request(app)
-                  .get('/static-three/bundle.js')
-                  .expect(200, (thirdError) => {
-                    if (thirdError) {
-                      return done(thirdError);
-                    }
-
-                    expect(getLogsPlugin.logs).toMatchSnapshot();
-
-                    return done();
-                  });
-              });
-          });
-      });
-    });
-
-    describe('should respect the "stats" option with the "{ all: false, entrypoints: true }" value from the configuration', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler({
-          ...webpackConfig,
-          stats: { all: false, entrypoints: true },
-        });
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request to the bundle file', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            expect(getLogsPlugin.logs).toMatchSnapshot();
-
-            return done();
-          });
-      });
-    });
-
-    describe('should respect the "stats" option from the configuration in multi-compiler mode', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackMultiWarningConfig);
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request to bundle files', (done) => {
-        request(app)
-          .get('/static-one/bundle.js')
-          .expect(200, (firstError) => {
-            if (firstError) {
-              return done(firstError);
-            }
-
-            return request(app)
-              .get('/static-two/bundle.js')
-              .expect(200, (secondError) => {
-                if (secondError) {
-                  return done(secondError);
-                }
-
-                expect(getLogsPlugin.logs).toMatchSnapshot();
-
-                return done();
-              });
-          });
-      });
-    });
-
-    describe('should respect the "stats" option from the configuration in multi-compiler mode and use the "name" option', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler(
-          webpackOneErrorOneWarningOneSuccessWithNamesConfig
-        );
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for "GET" requests to bundle files', (done) => {
-        request(app)
-          .get('/static-one/bundle.js')
-          .expect(200, (firstError) => {
-            if (firstError) {
-              return done(firstError);
-            }
-
-            return request(app)
-              .get('/static-two/bundle.js')
-              .expect(200, (secondError) => {
-                if (secondError) {
-                  return done(secondError);
-                }
-
-                return request(app)
-                  .get('/static-three/bundle.js')
-                  .expect(200, (thirdError) => {
-                    if (thirdError) {
-                      return done(thirdError);
-                    }
-
-                    expect(getLogsPlugin.logs).toMatchSnapshot();
-
-                    return done();
-                  });
-              });
-          });
-      });
-    });
-
-    describe('should throw an error on "run" when we watching', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackConfig);
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should logging an error', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            return compiler.run((runError) => {
-              expect(() => {
-                throw runError;
-              }).toThrowErrorMatchingSnapshot();
-
-              done();
-            });
-          });
-      });
-    });
-
-    describe('should throw an error on "watch" when we watching', () => {
-      let compiler;
-      let getLogsPlugin;
-
-      beforeAll((done) => {
-        compiler = getCompiler(webpackConfig);
-
-        getLogsPlugin = new GetLogsPlugin();
-        getLogsPlugin.apply(compiler);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should logging an error', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            return compiler.watch({}, (watchError) => {
-              expect(() => {
-                throw watchError;
-              }).toThrowErrorMatchingSnapshot();
-
-              done();
-            });
-          });
-      });
-    });
-
-    describe('should handle an earlier request if a change happened while compiling', () => {
-      beforeAll((done) => {
-        const compiler = getCompiler(webpackConfig);
-
-        instance = middleware(compiler);
-
-        let invalidated = false;
-
-        compiler.hooks.done.tap('Invalidated', () => {
-          if (!invalidated) {
-            instance.invalidate();
-
-            invalidated = true;
-          }
-        });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request to the bundle file', (done) => {
-        request(app).get('/bundle.js').expect(200, done);
-      });
-    });
-  });
-
-  describe('mimeTypes option', () => {
-    describe('should set the correct value for "Content-Type" header to known MIME type', () => {
-      beforeAll((done) => {
-        const outputPath = path.resolve(__dirname, './outputs/basic');
-        const compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: outputPath,
-          },
-        });
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-
-        instance.context.outputFileSystem.mkdirSync(outputPath, {
-          recursive: true,
-        });
-        instance.context.outputFileSystem.writeFileSync(
-          path.resolve(outputPath, 'file.html'),
-          'welcome'
-        );
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request to "file.html"', (done) => {
-        request(app)
-          .get('/file.html')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(200, 'welcome', done);
-      });
-    });
-
-    describe('should set the correct value for "Content-Type" header to unknown MIME type', () => {
-      beforeAll((done) => {
-        const outputPath = path.resolve(__dirname, './outputs/basic');
-        const compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: outputPath,
-          },
-        });
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-
-        instance.context.outputFileSystem.mkdirSync(outputPath, {
-          recursive: true,
-        });
-        instance.context.outputFileSystem.writeFileSync(
-          path.resolve(outputPath, 'file.phtml'),
-          'welcome'
-        );
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request to "file.phtml"', (done) => {
-        request(app)
-          .get('/file.phtml')
-          .expect('Content-Type', 'application/octet-stream')
-          .expect(200, done);
-      });
-    });
-
-    describe('should set the correct value for "Content-Type" header to specified MIME type', () => {
-      beforeAll((done) => {
-        const outputPath = path.resolve(__dirname, './outputs/basic');
-        const compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: outputPath,
-          },
-        });
-
-        instance = middleware(compiler, {
-          mimeTypes: {
-            phtml: 'text/html',
-          },
-        });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-
-        instance.context.outputFileSystem.mkdirSync(outputPath, {
-          recursive: true,
-        });
-        instance.context.outputFileSystem.writeFileSync(
-          path.resolve(outputPath, 'file.phtml'),
-          'welcome'
-        );
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request "file.phtml"', (done) => {
-        request(app)
-          .get('/file.phtml')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(200, 'welcome', done);
-      });
-    });
-
-    describe('should override value for "Content-Type" header for known MIME type', () => {
-      beforeAll((done) => {
-        const outputPath = path.resolve(__dirname, './outputs/basic');
-        const compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: outputPath,
-          },
-        });
-
-        instance = middleware(compiler, {
-          mimeTypes: {
-            jpg: 'application/octet-stream',
-          },
-        });
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-
-        instance.context.outputFileSystem.mkdirSync(outputPath, {
-          recursive: true,
-        });
-        instance.context.outputFileSystem.writeFileSync(
-          path.resolve(outputPath, 'file.jpg'),
-          'welcome'
-        );
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request "file.jpg"', (done) => {
-        request(app)
-          .get('/file.jpg')
-          .expect('Content-Type', /application\/octet-stream/)
-          .expect(200, done);
-      });
-    });
-
-    describe('should not set "Content-Type" header for route not from outputFileSystem', () => {
-      beforeAll((done) => {
-        const outputPath = path.resolve(__dirname, './outputs/basic');
-        const compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: outputPath,
-          },
-        });
-
-        instance = middleware(compiler, {
-          mimeTypes: {
-            jpg: 'application/octet-stream',
-          },
-        });
-
-        app = express();
-        app.use(instance);
-
-        app.get('/file.jpg', (req, res) => {
-          res.send('welcome');
-        });
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should return the "200" code for the "GET" request "file.jpg" with default content type', (done) => {
-        request(app)
-          .get('/file.jpg')
-          .expect('Content-Type', /text\/html/)
-          .expect(200, done);
-      });
-    });
-  });
 
   describe('watchOptions option', () => {
     describe('should work without value', () => {
